chore: Update setup.cfg and requirements
This commit updates the setup.cfg file to correct a typo in the description and update the Python version requirement to >=3.9. It also adds two new dependencies, matplotlib and tifffile, to the install_requires section.

The commit also removes the Two-Channels-Correlation_notebook.ipynb file from the notebooks directory.

Lastly, it adds a console script entry point in the options.entry_points section of setup.cfg.

This commit is a chore as it includes changes related to project configuration and dependencies.